requirements.txt
python-dotenv==0.15.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
eventlet==0.30.2
SQLAlchemy==1.3.22
requests==2.25.1
boto3==1.16.58
sendgrid==6.5.0
cryptography==3.3.1
psycopg2-binary==2.8.6
Flask-SocketIO==5.0.1
celery==5.0.5
redis==3.5.3
twilio==6.63.1
authlib==0.15.5
flask_dance==5.1.0
google-api-python-client==2.31.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
oauth2client==4.1.3

